LPS-53617 Java code: move web classes
