+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/notification/journal-user-notification-definitions.xml
