+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/portlet/action/ExportArticleActionCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.journal.action;
+package com.liferay.journal.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseActionCommand;
 import com.liferay.portal.util.PortalUtil;
