+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/portlet/route/journal-friendly-url-routes.xml
