+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/portlet/template/dependencies/link-to-page.ftl
