+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/search/FeedSearchTerms.java
@@ -12,9 +12,8 @@
  * details.
  */
 
-package com.liferay.portlet.journal.search;
+package com.liferay.journal.web.search;
 
-import com.liferay.portal.kernel.dao.search.DisplayTerms;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.WebKeys;
@@ -24,17 +23,9 @@ import javax.portlet.PortletRequest;
 /**
  * @author Raymond Augé
  */
-public class FeedDisplayTerms extends DisplayTerms {
+public class FeedSearchTerms extends FeedDisplayTerms {
 
-	public static final String DESCRIPTION = "description";
-
-	public static final String FEED_ID = "searchFeedId";
-
-	public static final String GROUP_ID = "groupId";
-
-	public static final String NAME = "name";
-
-	public FeedDisplayTerms(PortletRequest portletRequest) {
+	public FeedSearchTerms(PortletRequest portletRequest) {
 		super(portletRequest);
 
 		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
@@ -46,26 +37,5 @@ public class FeedDisplayTerms extends DisplayTerms {
 			portletRequest, GROUP_ID, themeDisplay.getScopeGroupId());
 		name = ParamUtil.getString(portletRequest, NAME);
 	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public String getFeedId() {
-		return feedId;
-	}
-
-	public long getGroupId() {
-		return groupId;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	protected String description;
-	protected String feedId;
-	protected long groupId;
-	protected String name;
 
 }
